Refactor tasks.rb
Extract internal TaskTiming class
